sendId -> sendid, since lowercase is used elsewhere
See bug exposed here (and when the "I" is replaced with an "i" the problem goes away).

SimpleInterpreter.prototype._send looks for "options.sendid" not "options.sendId" which meant that the id attribute was ignored altogether, making them impossible to cancel.  Cancelling events now work.  How goes it with "scion-ng" ?