Removed the setTimeout in beginAnimation on merge.
Now we only use it for layout. This is because it was causing a bug in
rendering updates in scxml-preview, where we would see a single frame of
the animation end state when an element was added or removed to/from the
graph.

However, this causes a regression in chrome for test2 -> test3 -> test4.
The second part of the spline disappears when the animation completes. I
think this is because of a bug, and has to do with stroke-dashoffset
property being set. I will investigate.