Carve out a code path to defer compilation of generated code...
So that we can generate a big module string which can be generated at
build-time and saved to disk. This is important for two reasons:

1. This allows the browser to avoid using scion compiler at runtime.
Compiled SCXML-to-js modules are much smaller runtime dependency over
the wire than the SCION compiler.

2. Node.js --inspect does not work for code evaluated at runtime using
eval or node vm module. See issue https://github.com/nodejs/node/issues/8369.
So to use chrome dev tools and source maps to debug SCXML, it is
currently necessary to use ahead-of-time compilation to generate a
module file as a workaround.