undo change to more-parallel/test2
changed the test script for more-parallel/test2
create a new test based on changed more-parallel/test2
write a note about the change in semantics:

Change in pre-emption logic between SCION-CORE 1.*.* semantics and SCION
3.*.* ("SCION 3 semantics"), and irp-tests branch, which implements the
logic defined in the Algorithm for SCXML Interpretation ("SCXML
semantics") [0].

In test more-parallel/test2, transitions a -> b and b1 -> b2 are both
enabled by event t.  Because the transition scopes of a -> b is an
ancestor of the transition scope of b1 -> b2, the the transitions scopes
 conflict. In SCION 3 semantics, transition b1 -> b2 preempts (has
higher priority than) transition a -> b; whereas in SCXML semantics, the
reverse is true: transition a -> b preempts b1 -> b2. In SCION 3
semantics, this is because the logic is: compare the transitions first
by depth (deepest first); and if there is a tie, then select the
transition by document order (earlier first). In SCXML semantics the
logic: "If multiple matching transitions are present, take
the first in document order. If none are present, search in the state's
ancestors in ancestry order until one is found." [1]

[0] https://www.w3.org/TR/scxml/#AlgorithmforSCXMLInterpretation
[1] https://www.w3.org/TR/scxml/#selectTransitions