Ugly workaround for issue nodejs/node#7593
Copy executionContext into global variable. This is ugly, because it
means that identically-named properties of different execution contexts
can step on each other in subsequent calls to model.prepare. This is
ok-ish, because SCION support for node --inspect is a development-only
feature. This will not affect production uses of SCION.

One day I may patch require to accept an executionContext argument, and
use vm.runInContext inside of Module._compile.