Extended the SCHVIZ component API to accept the following props:
* pathToSCXML
* urlToSCXML
* scxmlDocumentString
* modelFactory
* scjson

This mirrors the existing SCION model parser.

Modified the test harness to test these different API props. Ad hoc
tests indicate that everything is working.

Added SCION as a dependency to SCHVIZ. It might be better to break this
out later, if we want to use SCHVIZ as a lightweight SVG renderer for KGraph,
without supporting SCXML.