Some fixes related to key handling behavior:
Esc was not being captured with jquery keypress event handler. So
switched to keydown. event.keyCode and event.charCode properties
have been deprecated, so switched to event.key event. Updated in the
references in editor-behavior.scxml file.

Some features still not fully working. For example, macros: register
selection fails (e.g. command "qta" to start recording macro into
register "a").

* Also, updated to use latest and greated SCION with debug tooling support.
* Changed SVG file to HTML file. Pure SVG browser context is
incompatible with vm-browserify2 module. This is beacause it attempts to
create an HTML iframe and insert it into the SVG page. In the future,
could be fixed, if support for pure SVG context is needed.